[refactor] dto 폴더구조 request와 response로 분리
