+++ src/main/java/com/afternote/domain/timeletter/dto/request/TimeLetterDeleteRequest.java
@@ -1,4 +1,4 @@
-package com.afternote.domain.timeletter.dto;
+package com.afternote.domain.timeletter.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotEmpty;
